Added start screen to spawn players

+++ PartyGame/Source/PartyGame/CameraHandler.h
@@ -6,27 +6,26 @@
 #include "CameraHandler.generated.h"
 
 UCLASS()
-class PARTYGAME_API ACameraHandler : public AActor
-{
+class PARTYGAME_API ACameraHandler : public AActor {
 	GENERATED_BODY()
-	
-public:	
+
+public:
 	// Sets default values for this actor's properties
 	ACameraHandler();
 
 	UPROPERTY(EditAnywhere)
-	int lerpSpeed = 10;
+		float lerpSpeed = 10;
 	UPROPERTY(EditAnywhere)
-	int SpringArmBase = 300;
+		float SpringArmBase = 300;
 	UPROPERTY(EditAnywhere)
-	int SpringArmMultiplier = 2;
+		float SpringArmMultiplier = 2;
 	// Called when the game starts or when spawned
 	virtual void BeginPlay() override;
-	
+
 	// Called every frame
-	virtual void Tick( float DeltaSeconds ) override;
+	virtual void Tick(float DeltaSeconds) override;
 
 	UFUNCTION(BlueprintCallable, Category = "CameraHandler")
 		void LerpTowardsAverage();
-	
+
 };
